new allow restore selector
